serving/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pandas